ohlc-exporter/requirements.txt
prometheus_client
websocket-client